section4 complete

+++ go-stripe/cmd/web/render.go
@@ -29,7 +29,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	return fmt.Sprintf("$%.2f", float32(n)/100)
+	return fmt.Sprintf("$%.2f", float32(n)/float32(100))
 }
 
 //go:embed templates
